+++ database/migrations/2023_01_19_173527_create_permanences_table.php
@@ -13,12 +13,14 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('permanences', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->year('year_permanence');
-            $table->timestamps();
-        });
+        if(!Schema::hasTable('permanences')){
+            Schema::create('permanences', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                $table->year('year_permanence');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
