+++ database/migrations/2023_05_10_145237_companies_address_table.php
@@ -13,19 +13,21 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('companies_address', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('companie_id')->constrained()->onDelete('cascade');
-            $table->string('via')->nullable();
-            $table->string('direccion')->nullable();
-            $table->string('piso')->nullable();
-            $table->integer('cp')->nullable();
-            $table->string('ciudad')->nullable();
-            $table->string('provincia')->nullable();
-            $table->string('coord-x')->nullable();
-            $table->string('coord-y')->nullable();
-            $table->timestamps();
-        });
+        if(!Schema::hasTable('companies_address')){
+            Schema::create('companies_address', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('companie_id')->constrained()->onDelete('cascade');
+                $table->string('via')->nullable();
+                $table->string('direccion')->nullable();
+                $table->string('piso')->nullable();
+                $table->integer('cp')->nullable();
+                $table->string('ciudad')->nullable();
+                $table->string('provincia')->nullable();
+                $table->string('coord-x')->nullable();
+                $table->string('coord-y')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
