+++ database/migrations/2023_06_01_212619_add_user_to_notifications.php
@@ -13,10 +13,12 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::table('notifications', function (Blueprint $table) {
-            $table->unsignedBigInteger('id_user')->nullable();
-            $table->foreign('id_user')->references('id')->on('users')->onDelete('cascade');
-        });
+        if (!Schema::hasColumn('notifications', 'id_user')){
+            Schema::table('notifications', function (Blueprint $table) {
+                $table->unsignedBigInteger('id_user')->nullable();
+                $table->foreign('id_user')->references('id')->on('users')->onDelete('cascade');
+            });
+        }
     }
 
     /**
